use latest houston and introduce a flag to keep files on disk

+++ lib/stars/runner.go
@@ -26,11 +26,12 @@ const (
 )
 
 type RunnerOptions struct {
-	ExecutableDir   string `long:"stars-executable-dir" env:"STARS_EXECUTABLE_DIR" ini-name:"stars_executable_dir" description:"directory that will be used to put stars.exe" default:"~/.local/share/neper/stars"`
-	SaveDir         string `long:"stars-save-dir" env:"STARS_SAVE_DIR" ini-name:"stars_save_dir" description:"directory that will be used as a base to create all savegame dirs" default:"~/.local/share/neper/saves"`
-	WinePrefix      string `long:"wine-prefix" env:"WINE_PREFIX" ini-name:"wine_prefix" description:"wine prefix to use for running wine apps" default:"~/.wine"`
-	CommandsTimeout int    `long:"wine-commands-timeout" env:"WINE_COMMANDS_TIMEOUT" ini-name:"wine_commands_timeout" description:"time in seconds after which the wine process is considered unresponsive and killed" default:"30"`
-	DisplayNumber   int    `long:"display" env:"DISPLAY" ini-name:"display" description:"the display number to provide for wine commands" default:"99"`
+	ExecutableDir    string `long:"stars-executable-dir" env:"STARS_EXECUTABLE_DIR" ini-name:"stars_executable_dir" description:"directory that will be used to put stars.exe" default:"~/.local/share/neper/stars"`
+	SaveDir          string `long:"stars-save-dir" env:"STARS_SAVE_DIR" ini-name:"stars_save_dir" description:"directory that will be used as a base to create all savegame dirs" default:"~/.local/share/neper/saves"`
+	WinePrefix       string `long:"wine-prefix" env:"WINE_PREFIX" ini-name:"wine_prefix" description:"wine prefix to use for running wine apps" default:"~/.wine"`
+	CommandsTimeout  int    `long:"wine-commands-timeout" env:"WINE_COMMANDS_TIMEOUT" ini-name:"wine_commands_timeout" description:"time in seconds after which the wine process is considered unresponsive and killed" default:"30"`
+	DisplayNumber    int    `long:"display" env:"DISPLAY" ini-name:"display" description:"the display number to provide for wine commands" default:"99"`
+	PreserveSaveFiles bool  `long:"preserve-save-files" env:"PRESERVE_SAVE_FILES" ini-name:"preserve_save_files" description:"preserve session save files on disk for debugging (do not cleanup after game creation/turn generation)"`
 }
 
 // NewRunnerOptions ...
@@ -40,13 +41,14 @@ func NewRunnerOptions() *RunnerOptions {
 }
 
 type Runner struct {
-	log           *zerolog.Logger
-	ExecutableDir string // will be mapped to x: for executables
-	SaveDir       string // will be mapped to s: for saves
-	DisplayNumber int
-	prefix        *wine.Prefix    // wine prefix manager
-	xvfbProcess   *cmd.Cmd        // the xvfbProcess we launch at startup
-	xvfbStatusChan <-chan cmd.Status
+	log               *zerolog.Logger
+	ExecutableDir     string // will be mapped to x: for executables
+	SaveDir           string // will be mapped to s: for saves
+	DisplayNumber     int
+	preserveSaveFiles bool              // if true, don't cleanup session directories after operations
+	prefix            *wine.Prefix      // wine prefix manager
+	xvfbProcess       *cmd.Cmd          // the xvfbProcess we launch at startup
+	xvfbStatusChan    <-chan cmd.Status
 }
 
 func expand(input string) (string, error) {
@@ -78,11 +80,12 @@ func NewRunner(log *zerolog.Logger, opts *RunnerOptions) (*Runner, error) {
 	}
 
 	return &Runner{
-		log:           log,
-		ExecutableDir: absExecutableDir,
-		SaveDir:       absSaveDir,
-		DisplayNumber: opts.DisplayNumber,
-		prefix:        prefix,
+		log:               log,
+		ExecutableDir:     absExecutableDir,
+		SaveDir:           absSaveDir,
+		DisplayNumber:     opts.DisplayNumber,
+		preserveSaveFiles: opts.PreserveSaveFiles,
+		prefix:            prefix,
 	}, nil
 }
 
@@ -281,8 +284,13 @@ func (r *Runner) Prefix() *wine.Prefix {
 
 // CleanupSessionDir removes the session directory and all its contents.
 // This should be called after successfully saving game files to the database.
+// If preserveSaveFiles is enabled, the directory is kept for debugging purposes.
 func (r *Runner) CleanupSessionDir(sessionID string) error {
 	sessionDir := r.localSessionSaveDir(sessionID)
+	if r.preserveSaveFiles {
+		r.log.Debug().Str("sessionID", sessionID).Str("dir", sessionDir).Msg("preserving session directory for debugging")
+		return nil
+	}
 	if err := os.RemoveAll(sessionDir); err != nil {
 		r.log.Err(err).Str("sessionID", sessionID).Str("dir", sessionDir).Msg("failed to cleanup session directory")
 		return err
